Move endpoint entities to CountryInfo object
